Planner AI/requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib